+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 - Bugfix disable track output when restarting ([#134](https://github.com/OpenRemise/Firmware/issues/134))
+- Bugfix settings reset must always work ([#135](https://github.com/OpenRemise/Firmware/issues/135))
 - Bugfix Z21 `LAN_X_SET_STOP` and `LAN_X_SET_LOCO_E_STOP` ([#136](https://github.com/OpenRemise/Firmware/issues/136))
 
 ## 0.7.0